+++ Lab6/schemas/finance/finance_schema_update.py
@@ -6,5 +6,5 @@ from pydantic import BaseModel
 class FinanceSchemaUpdate(BaseModel):
     id: int
     action: int
-    sum: float
+    sum: int
     time: Optional[datetime]
